ButtonEventsProduct renamed to CreateNewSparePartButtonEvent code extracted to SparePartFactory class
